[add] user authentication

+++ JssProject/main/views.py
@@ -2,11 +2,15 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .forms import JssForm
 from .models import Jasoseol
 from django.http import Http404
+from django.core.exceptions import PermissionDenied
+from django.contrib.auth.decorators import login_required
 # Create your views here.
 def index(request):
+    #all_jss는 '모든' user가 쓴 글에 관한 정보
     all_jss = Jasoseol.objects.all()
     return render(request, 'index.html', {'all_jss' : all_jss})
 
+@login_required(login_url='/login/')
 def create(request):
     #POST에 관련한 함수도 create에서 다 처리를 하고 싶기 때문에 여기다 추가
     if request.method == "POST":
@@ -15,6 +19,10 @@ def create(request):
         filled_form = JssForm(request.POST)
         #is_valid라는 함수는 들어온 데이터가 문제가 없는지 확인 후 문제가 없다면 다음으로 넘어감
         if filled_form.is_valid(): 
+            #임시 form 생성
+            #commit 이라는 속성은 잠깐 코드를 지연시키고 그 안에서 원하는 행위들을 해줄 수 있음
+            temp_form = filled_form.save(commit=False)
+            temp_form.author = request.user
             filled_form.save()
             return redirect('index')
     jss_form = JssForm()
@@ -29,10 +37,15 @@ def detail(request, jss_id):
     return render(request, 'detail.html', {'my_jss' : my_jss})
 
 def delete(reqeust, jss_id):
-    my_jss =Jasoseol.objects.get(pk=jss_id)
-    my_jss.delete()
-    return redirect('index')
+    #my_jss는 '특정'유저가 쓴 글에 관한 정보
+    my_jss = Jasoseol.objects.get(pk=jss_id)
+    if reqeust.user == my_jss.author :
+        my_jss.delete()
+        return redirect('index')
 
+    raise PermissionDenied
+    
+    
 def update(request, jss_id):
     my_jss =Jasoseol.objects.get(pk=jss_id)
     #modelform을 사용하겠다고 선언
